add new files and methods

+++ src/main/java/com/appium/manager/AppiumServerManager.java
@@ -112,7 +112,7 @@ public class AppiumServerManager {
             List<Device> devices = DeviceAllocationManager.stfService.getDevices().getDevices();
             devices.forEach(device -> {
                 if (device.isPresent() && device.getOwner() != null && device.isUsing() &&
-                        device.getUsage().equalsIgnoreCase("automation")) {
+                        (device.getUsage()!=null &&  device.getUsage().equalsIgnoreCase("automation"))) {
                     DeviceAllocationManager.stfService.deleteDeviceBySerial(device.getSerial());
                 }
             });
